int changed to float

+++ webapp/payments/paypal.py
@@ -24,11 +24,11 @@ def create_paypal_order(currency, price, quantity, order_id, receiver, product_n
                     {
                         "amount": {
                             "currency_code": currency,
-                            "value": str(int(quantity) * int(price)),
+                            "value": str(float(quantity) * float(price)),
                             "breakdown": {
                                 "item_total" : {
                                      "currency_code" : currency,
-                                        "value" : str(int(quantity) * int(price))
+                                        "value" : str(float(quantity) * float(price))
                                 }
                             }
                         },
